services/api-service/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy[asyncio]==2.0.35",
    "asyncpg==0.29.0",
    "alembic==1.13.3",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "redis[asyncio]==5.1.1",
    "confluent-kafka==2.5.3",
    "fastavro==1.9.7",
    "pydantic-settings==2.5.2",
    "structlog==24.4.0",
    "prometheus-client==0.21.0",
    "tenacity==9.0.0",
    "httpx==0.27.2",
    "boto3==1.35.36",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "pytest-cov==5.0.0",
    "anyio==4.6.0",
    "aiosqlite==0.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/api_service"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
